backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Acaceta product"
authors = ["ethanelasky <elaskyethan@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
httpx = "<0.28.0"
python = ">=3.12,<3.13"
Flask = "^3.0.3"
Flask-Cors = "^4.0.1"
flask-mongoengine-3 = "^1.0.9"
opensearch-py = "^2.6.0"
boto3 = "^1.34.147"
Flask-Mail = "^0.10.0"
Werkzeug = "^3.0.3"
mongoengine = "^0.29.0"
uuid = "^1.30"
APScheduler = "^3.10.4"
pymongo = "^4.8.0"
PyMuPDF = "1.24.8"
langchain = "^0.2.11"
langchain-core = "^0.2.23"
langchain-text-splitters = "^0.2.2"
loguru = "^0.7.2"
requests = "^2.32.3"
reportlab = "^4.2.2"
pypdf = "^4.3.1"
ratelimit = "^2.2.1"
urllib3 = "^1.2.6"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.1"
pyjwt = "^2.8.0"
flask-jwt-extended = "^4.6.0"
replicate = "^0.29.0"
environs = "^11.0.0"
matplotlib = "^3.9.1"
structlog = "^24.4.0"
asyncio = "^3.4.3"
nest-asyncio = "^1.6.0"
pypdf2 = "^3.0.1"
jsonwebtoken = "^1.0.0"
pdf2image = "^1.17.0"
poppler-utils = "^0.1.0"
fuzzywuzzy = "^0.18.0"
nltk = "^3.8.1"
cohere = "^5.8.1"
python-levenshtein = "^0.25.1"
prometheus-client = "^0.20.0"
pytz = "^2024.1"
elasticsearch = "^8.15.0"
stripe = "^10.11.0"
ngrok = "^1.4.0"
groq = "^0.12.0"
anthropic = "0.39.0"
openai = "^1.48.0"
flask-limiter = "^3.8.0"
pytest = "^8.3.3"
mock = "^5.1.0"
pytest-mock = "^3.14.0"
playwright-stealth = "^1.0.6"
gunicorn = "^23.0.0"
deepeval = "^1.4.7"
pytest-xdist = "^3.6.1"
pre-commit = "^4.0.1"
pytest-asyncio = "^0.24.0"
pytest-env = "^1.1.5"
types-requests = "2.31.0.6"
vcrpy = "^6.0.2"
seaborn = "^0.13.2"
voyageai = "^0.3.2"
uvicorn = "^0.32.1"
gevent = "^24.11.1"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
